order_summary_agent/agent.py
from google.adk.agents import LlmAgent
from google.adk.models.lite_llm import LiteLlm


local_model = LiteLlm(
    model="ollama_chat/qwen2.5:7b-instruct"
)


order_summary_agent = LlmAgent(
    name="order_summary_agent",
    description="Order summary agent",
    model=local_model,
    instruction="""
SYSTEM ROLE:
You are an ORDER SUMMARY AGENT.

RULES:
- Read ONLY from SESSION STATE
- Do NOT invent data
- Do NOT ask questions
- Do NOT greet

OUTPUT FORMAT (STRICT):

Order Summary
-------------------
Name: {name}
Email: {email}
Mobile: {mobile}

Item: {item}
Quantity: {quantity}
Price: ₹{price}

Shipping Address:
{shipping_address}

If any field is missing, write: Not provided
"""
)
